+++ dtale/dash_application/components/_imports_.py
@@ -1,0 +1,5 @@
+from .Wordcloud import Wordcloud
+
+__all__ = [
+    "Wordcloud"
+]