+++ static/dash/lib/index.js
@@ -1,0 +1,3 @@
+import Wordcloud from "./components/Wordcloud.react";
+
+export { Wordcloud };